feat: implement error dto and business base #145

+++ backend/src/main/java/ch/puzzle/pcts/service/business/CertificateTypeBusinessService.java
@@ -1,29 +1,34 @@
 package ch.puzzle.pcts.service.business;
 
+import static ch.puzzle.pcts.Constants.CERTIFICATE_TYPE;
+
 import ch.puzzle.pcts.exception.PCTSException;
 import ch.puzzle.pcts.model.certificatetype.CertificateType;
 import ch.puzzle.pcts.model.error.ErrorKey;
+import ch.puzzle.pcts.model.error.FieldKey;
+import ch.puzzle.pcts.repository.CertificateTypeRepository;
 import ch.puzzle.pcts.service.persistence.CertificateTypePersistenceService;
 import ch.puzzle.pcts.service.validation.CertificateTypeValidationService;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CertificateTypeBusinessService {
-    private final CertificateTypeValidationService validationService;
-    private final CertificateTypePersistenceService persistenceService;
+public class CertificateTypeBusinessService
+        extends
+            BusinessBase<CertificateType, CertificateTypeRepository, CertificateTypeValidationService, CertificateTypePersistenceService> {
     private final TagBusinessService tagBusinessService;
 
     public CertificateTypeBusinessService(CertificateTypeValidationService certificateTypeValidationService,
                                           CertificateTypePersistenceService certificateTypePersistenceService,
                                           TagBusinessService tagBusinessService) {
-        this.validationService = certificateTypeValidationService;
-        this.persistenceService = certificateTypePersistenceService;
+        super(certificateTypeValidationService, certificateTypePersistenceService);
         this.tagBusinessService = tagBusinessService;
     }
 
+    @Override
     public CertificateType create(CertificateType certificateType) {
         validationService.validateOnCreate(certificateType);
 
@@ -32,6 +37,7 @@ public class CertificateTypeBusinessService {
         return persistenceService.save(certificateType);
     }
 
+    @Override
     public CertificateType getById(Long id) {
         validationService.validateOnGetById(id);
 
@@ -43,8 +49,14 @@ public class CertificateTypeBusinessService {
             return certificateType;
         } else {
             throw new PCTSException(HttpStatus.NOT_FOUND,
-                                    "Certificate type with id: " + id + " does not exist.",
-                                    ErrorKey.NOT_FOUND);
+                                    ErrorKey.NOT_FOUND,
+                                    Map
+                                            .of(FieldKey.ENTITY,
+                                                entityName(),
+                                                FieldKey.FIELD,
+                                                "id",
+                                                FieldKey.IS,
+                                                id.toString()));
         }
     }
 
@@ -52,24 +64,56 @@ public class CertificateTypeBusinessService {
         return persistenceService.getAllCertificateTypes();
     }
 
+    @Override
     public CertificateType update(Long id, CertificateType certificateType) {
         validationService.validateOnUpdate(id, certificateType);
 
         certificateType.setTags(tagBusinessService.resolveTags(certificateType.getTags()));
-        certificateType.setId(id);
+        if (persistenceService.getById(id).isPresent()) {
 
-        CertificateType result = persistenceService.save(certificateType);
-        tagBusinessService.deleteUnusedTags();
+            certificateType.setId(id);
 
-        return result;
+            CertificateType result = persistenceService.save(certificateType);
+            tagBusinessService.deleteUnusedTags();
+
+            return result;
+        } else {
+            throw new PCTSException(HttpStatus.NOT_FOUND,
+                                    ErrorKey.NOT_FOUND,
+                                    Map
+                                            .of(FieldKey.ENTITY,
+                                                entityName(),
+                                                FieldKey.FIELD,
+                                                "id",
+                                                FieldKey.IS,
+                                                id.toString()));
+        }
     }
 
+    @Override
     public void delete(Long id) {
         validationService.validateOnDelete(id);
 
-        persistenceService.delete(id);
+        if (persistenceService.getById(id).isPresent()) {
+            persistenceService.delete(id);
 
-        tagBusinessService.deleteUnusedTags();
+            tagBusinessService.deleteUnusedTags();
+        } else {
+            throw new PCTSException(HttpStatus.NOT_FOUND,
+                                    ErrorKey.NOT_FOUND,
+                                    Map
+                                            .of(FieldKey.ENTITY,
+                                                entityName(),
+                                                FieldKey.FIELD,
+                                                "id",
+                                                FieldKey.IS,
+                                                id.toString()));
+        }
+    }
+
+    @Override
+    protected String entityName() {
+        return CERTIFICATE_TYPE;
     }
 
 }
